Fix EINTR while doing system calls
This commit fixes the problem with socket.accept() and socket.recvfrom() raising
Interrupted System Call exception.  The exception is handled by simplying
retrying the respective system calls.  The reason why this happens to be is that
we are using multiprocessing.Pipe to communicate with the Firewall handling
process and that causes signals to be raised somehow.