Fix concurrency issues while adding a request into the set of currently active port open requests.
The hash for a request is calculated from the IP address, IP address type, port,
transport.  This hash is then checked in a set to determine whether the same
request is currently active.  If not, a new port open thread is opened and this
hash is added to the set.

Earlier, the thread opening the port is adding this hash and this leads to race
conditions where two threads can be started to add the same request hash.  This
is because the lock used to maintain synchronization is released before the
thread has started.  This commit fixes this by moving the code adding the hash
from the thread to where the check is made.